+++ src/render/index.js
@@ -1,13 +1,19 @@
 'use strict';
 
+// default TTL is 60 minutes
+var DEFAULT_TTL = 1000 * 60 * 60;
+
 var loader = require('./loader');
 var render = require('./render');
 var cache = require('./cache');
 
 var pathPrefix = '';
 
-exports.config = function (path) {
+exports.config = function (path, cacheSize) {
 	pathPrefix = path;
+	if (cacheSize !== undefined) {
+		cache.setMemSize(cacheSize);
+	}
 };
 
 exports.setup = function (cb) {
@@ -18,6 +24,9 @@ exports.setup = function (cb) {
 };
 
 exports.render = function (path, vars, cacheTtl) {
+	if (cacheTtl === undefined) {
+		cacheTtl = DEFAULT_TTL;
+	}
 	if (cacheTtl) {
 		var res = cache.get(vars);
 		if (res) {